# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Qragy API",
    "description": "Self-hosted AI support chatbot with RAG, human-in-the-loop, and full admin panel.",
    "version": "1.0.0",
    "contact": {
      "name": "Qragy",
      "url": "https://github.com/mahsumaktas/qragy"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "Admin panel authentication token"
      }
    },
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/jobs/stats": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get job queue stats",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Status-based job counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "stats": {
                      "type": "object",
                      "properties": {
                        "pending": { "type": "integer" },
                        "running": { "type": "integer" },
                        "completed": { "type": "integer" },
                        "dead": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/jobs/dead": {
      "get": {
        "tags": ["Jobs"],
        "summary": "List dead jobs",
        "security": [{ "AdminToken": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 50 },
            "description": "Max number of dead jobs to return"
          }
        ],
        "responses": {
          "200": {
            "description": "List of dead jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "jobs": { "type": "array", "items": { "type": "object" } }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Purge all dead jobs",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "Number of purged jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "purged": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/jobs/{id}/retry": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Retry a dead job",
        "security": [{ "AdminToken": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": { "description": "Job requeued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } },
          "404": { "description": "Dead job not found" }
        }
      }
    }
  }
}